# docs/api/swagger.yaml
openapi: 3.0.3
info:
  title: AIPC Platform API
  description: 'AIPC Platform REST API for edge AI computing platform.


    Provides unified access to:

    - AI Runtime (model management, inference)

    - Event Bus (pub/sub messaging)

    - Device Control (camera, PTZ, GPIO)

    - App Manager (container lifecycle)

    - Stream (HLS video streaming)

    '
  version: 1.0.0
  contact:
    name: AIPC Platform Team
    email: support@camthink.ai
servers:
- url: http://localhost:8080/api/v1
  description: Development server
- url: http://{host}:8080/api/v1
  description: Production server
  variables:
    host:
      default: localhost
security:
- BearerAuth: []
...
paths:
...
  /events/publish:
    post:
      tags:
      - Event Bus
      summary: Publish an event
      operationId: publishEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - topic
              properties:
                topic:
                  type: string
                  example: app/detection/result
                payload:
                  type: object
                  additionalProperties: true
      responses:
        '200':
          description: Event published
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/APIResponse'
                - type: object
                  properties:
                    data:
                      type: object
                      description: Empty data
...
  /storage/mount:
    post:
      tags: [Storage]
      summary: Mount a block device to a destination
      operationId: mountDisk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device]
              properties:
                device:
                  type: string
                  example: /dev/sda1
                target:
                  type: string
                  example: /mnt/sda1
      responses:
        '200':
          description: Mount successful
  /storage/unmount:
    post:
      tags: [Storage]
      summary: Unmount a previously mounted disk
      operationId: unmountDisk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target]
              properties:
                target:
                  type: string
                  example: /mnt/sda1
      responses:
        '200':
          description: Unmount successful
  /storage/format:
    post:
      tags: [Storage]
      summary: Format a disk with a specific filesystem
      operationId: formatDisk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [device]
              properties:
                device:
                  type: string
                  example: /dev/sda1
                fstype:
                  type: string
                  example: ext4
                  enum: [ext4, vfat, fat32]
      responses:
        '200':
          description: Format successful

  # ==================== Media Config ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  schemas:
    APIResponse:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: Success
        data:
          type: object
          description: Response payload
        error:
          $ref: '#/components/schemas/ErrorDetail'
...
    ErrorDetail:
      type: object
      properties:
        type:
          type: string
        detail:
          type: string

    # ==================== Store Schemas ====================
...